requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
asgiref==3.5.2
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.0; python_version >= '3.6'
click-didyoumean==0.3.0; python_version < '4' and python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3; python_version >= '3.7'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-celery-beat==2.3.0
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-redis==5.2.0
django-timezone-field==5.0; python_version >= '3.7' and python_version < '4'
django==3.2.7
djangorestframework-simplejwt==5.1.0
djangorestframework==3.13.1
drf-errors-formatter==1.0.1
drf-spectacular==0.22.1
gunicorn==20.1.0
idna==3.3; python_version >= '3.5'
inflection==0.5.1
jsonschema==4.5.1
kombu==5.2.4
lxml==4.9.1
packaging==21.3; python_version >= '3.6'
prompt-toolkit==3.0.30; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
pyjwt==2.4.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.1
pyyaml==6.0
redis==4.3.4
requests==2.28.1
setuptools==64.0.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlparse==0.4.2
tzdata==2022.2; python_version >= '2'
uritemplate==4.1.1
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
